Added login prototype
Warning - to make this work, you have to add it to your htdocs in your XAMPP directory and run it locally, as well as importing the demo database. Most of the code is not mine - we will have to edit this heavily to make it part of the project. However, it allows us a starting point.